requirements.txt
behave==1.2.6
pytest==7.4.3
faker==19.12.0
pydantic==2.5.0
python-dotenv==1.0.0
tabulate==0.9.0
colorama==0.4.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0 